fixed bug

+++ src/main/java/ServingLayer.java
@@ -8,14 +8,14 @@ import java.util.Map;
 
 public class ServingLayer {
 
-    static public Map<String, Integer> getResults(String[] keywords) {
+    static public Map<String, Long> getResults(String[] keywords) {
         SentimentRepository repository = SentimentRepository.getInstance();
-        Map<String, Integer> map = new HashMap<String, Integer>();
+        Map<String, Long> map = new HashMap<String, Long>();
         for (int i = 0; i < keywords.length; i++) {
             for (int j = -1; j <= 1; j++) {
                 String sentiment = Integer.toString(j);
                 String keyword_sentiment = keywords[i] + " | " + sentiment;
-                int count = repository.selectCountFromKey("fasttable", keywords[i], j);
+                long count = repository.selectCountFromKey("fasttable", keywords[i], j);
                 map.put(keyword_sentiment, count);
             }
         }
@@ -24,9 +24,9 @@ public class ServingLayer {
             for (int j = -1; j <= 1; j++) {
                 String sentiment = Integer.toString(j);
                 String keyword_sentiment = keywords[i] + " | " + sentiment;
-                int count = repository.selectCountFromKey("batchtable", keywords[i], j);
-                int val = map.get(keyword_sentiment);
-                int newVal = val + count;
+                long count = repository.selectCountFromKey("batchtable", keywords[i], j);
+                long val = map.get(keyword_sentiment);
+                long newVal = val + count;
                 map.put(keyword_sentiment, newVal);
             }
         }
